abstract aggregation functions to separate file
