+++ epilepsy12/common_view_functions/__init__.py
@@ -1,3 +1,4 @@
 from .calculate_kpis import calculate_kpis, hospital_level_kpis, trust_level_kpis, national_level_kpis
-from .recalculate_form_generate_response import recalculate_form_generate_response, completed_fields, test_fields_update_audit_progress
+from .recalculate_form_generate_response import recalculate_form_generate_response, completed_fields, test_fields_update_audit_progress, trigger_client_event
 from .validate_form_update_model import validate_and_update_model
+from .aggregate_by import cases_aggregated_by_deprivation_score, cases_aggregated_by_ethnicity, cases_aggregated_by_sex
