+++ templates/epilepsy12/partials/hospital/individual_metrics.html
@@ -1,3 +1,4 @@
+
 <div class="ui container">
   <p>
     Within the clinical audit there are 12 'Performance Indicator' measures
